third_party/WebKit/LayoutTests/external/wpt/selection/isCollapsed-expected.txt

CONSOLE ERROR: line 22: The given range isn't in document.
CONSOLE ERROR: line 22: The given range isn't in document.
CONSOLE ERROR: line 22: The given range isn't in document.
CONSOLE ERROR: line 22: The given range isn't in document.
CONSOLE ERROR: line 22: The given range isn't in document.
CONSOLE ERROR: line 22: The given range isn't in document.
CONSOLE ERROR: line 22: The given range isn't in document.
CONSOLE ERROR: line 22: The given range isn't in document.
CONSOLE ERROR: line 22: The given range isn't in document.
This is a testharness.js-based test.
Found 59 tests; 43 PASS, 16 FAIL, 0 TIMEOUT, 0 NOTRUN.
PASS Empty selection 
PASS Range 0 [paras[0].firstChild, 0, paras[0].firstChild, 0] 
PASS Range 1 [paras[0].firstChild, 0, paras[0].firstChild, 1] 
PASS Range 2 [paras[0].firstChild, 2, paras[0].firstChild, 8] 
PASS Range 3 [paras[0].firstChild, 2, paras[0].firstChild, 9] 
PASS Range 4 [paras[1].firstChild, 0, paras[1].firstChild, 0] 
PASS Range 5 [paras[1].firstChild, 0, paras[1].firstChild, 1] 
PASS Range 6 [paras[1].firstChild, 2, paras[1].firstChild, 8] 
PASS Range 7 [paras[1].firstChild, 2, paras[1].firstChild, 9] 
PASS Range 8 [detachedPara1.firstChild, 0, detachedPara1.firstChild, 0] 
FAIL Range 9 [detachedPara1.firstChild, 0, detachedPara1.firstChild, 1] assert_equals: Value of isCollapsed expected false but got true
FAIL Range 10 [detachedPara1.firstChild, 2, detachedPara1.firstChild, 8] assert_equals: Value of isCollapsed expected false but got true
PASS Range 11 [foreignPara1.firstChild, 0, foreignPara1.firstChild, 0] 
FAIL Range 12 [foreignPara1.firstChild, 0, foreignPara1.firstChild, 1] assert_equals: Value of isCollapsed expected false but got true
FAIL Range 13 [foreignPara1.firstChild, 2, foreignPara1.firstChild, 8] assert_equals: Value of isCollapsed expected false but got true
PASS Range 14 [document.documentElement, 0, document.documentElement, 1] 
PASS Range 15 [document.documentElement, 0, document.documentElement, 2] 
PASS Range 16 [document.documentElement, 1, document.documentElement, 2] 
PASS Range 17 [document.head, 1, document.head, 1] 
PASS Range 18 [document.body, 0, document.body, 1] 
FAIL Range 19 [foreignDoc.documentElement, 0, foreignDoc.documentElement, 1] assert_equals: Value of isCollapsed expected false but got true
PASS Range 20 [foreignDoc.head, 1, foreignDoc.head, 1] 
PASS Range 21 [foreignDoc.body, 0, foreignDoc.body, 0] 
PASS Range 22 [paras[0], 0, paras[0], 0] 
PASS Range 23 [paras[0], 0, paras[0], 1] 
PASS Range 24 [detachedPara1, 0, detachedPara1, 0] 
FAIL Range 25 [detachedPara1, 0, detachedPara1, 1] assert_equals: Value of isCollapsed expected false but got true
PASS Range 26 [paras[0].firstChild, 0, paras[1].firstChild, 0] 
PASS Range 27 [paras[0].firstChild, 0, paras[1].firstChild, 8] 
PASS Range 28 [paras[0].firstChild, 3, paras[3], 1] 
PASS Range 29 [paras[0], 0, paras[0].firstChild, 7] 
PASS Range 30 [testDiv, 2, paras[4], 1] 
PASS Range 31 [testDiv, 1, paras[2].firstChild, 5] 
PASS Range 32 [document.documentElement, 1, document.body, 0] 
FAIL Range 33 [foreignDoc.documentElement, 1, foreignDoc.body, 0] assert_equals: Value of isCollapsed expected false but got true
PASS Range 34 [document, 0, document, 1] 
PASS Range 35 [document, 0, document, 2] 
PASS Range 36 [document, 1, document, 2] 
PASS Range 37 [testDiv, 0, comment, 5] 
PASS Range 38 [paras[2].firstChild, 4, comment, 2] 
PASS Range 39 [paras[3], 1, comment, 8] 
PASS Range 40 [foreignDoc, 0, foreignDoc, 0] 
FAIL Range 41 [foreignDoc, 1, foreignComment, 2] assert_equals: Value of isCollapsed expected false but got true
FAIL Range 42 [foreignDoc.body, 0, foreignTextNode, 36] assert_equals: Value of isCollapsed expected false but got true
PASS Range 43 [xmlDoc, 0, xmlDoc, 0] 
FAIL Range 44 [xmlDoc, 1, xmlComment, 0] assert_equals: Value of isCollapsed expected false but got true
FAIL Range 45 [detachedTextNode, 0, detachedTextNode, 8] assert_equals: Value of isCollapsed expected false but got true
PASS Range 46 [detachedForeignTextNode, 7, detachedForeignTextNode, 7] 
FAIL Range 47 [detachedForeignTextNode, 0, detachedForeignTextNode, 8] assert_equals: Value of isCollapsed expected false but got true
PASS Range 48 [detachedXmlTextNode, 7, detachedXmlTextNode, 7] 
FAIL Range 49 [detachedXmlTextNode, 0, detachedXmlTextNode, 8] assert_equals: Value of isCollapsed expected false but got true
FAIL Range 50 [detachedComment, 3, detachedComment, 4] assert_equals: Value of isCollapsed expected false but got true
PASS Range 51 [detachedComment, 5, detachedComment, 5] 
FAIL Range 52 [detachedForeignComment, 0, detachedForeignComment, 1] assert_equals: Value of isCollapsed expected false but got true
PASS Range 53 [detachedForeignComment, 4, detachedForeignComment, 4] 
FAIL Range 54 [detachedXmlComment, 2, detachedXmlComment, 6] assert_equals: Value of isCollapsed expected false but got true
PASS Range 55 [docfrag, 0, docfrag, 0] 
PASS Range 56 [foreignDocfrag, 0, foreignDocfrag, 0] 
PASS Range 57 [xmlDocfrag, 0, xmlDocfrag, 0] 
Harness: the test ran to completion.
